fix(e2e): move correction proposal tests into import-wizard.spec.ts to reuse proven mock infra
Standalone correction-proposal.spec.ts failed because the custom mock setup
couldn't correctly advance the wizard past the Processing step. The existing
import-wizard.spec.ts mock infrastructure (setupMockAPIs + navigateToReviewStep)
is proven to work, so the correction proposal tests are colocated here and add
only the correction-specific routes on top.
