+++ apps/pops-shell/e2e/import-wizard.spec.ts
@@ -1612,3 +1612,187 @@ test.describe.skip('Import Wizard - Error Recovery', () => {
     await expect(page.locator('select[name="amount"]')).toHaveValue('Amount');
   });
 });
+
+// ---------------------------------------------------------------------------
+// Correction Proposal Dialog
+// ---------------------------------------------------------------------------
+// Verifies that editing a rule field auto-reruns the preview so Apply stays
+// reachable without forcing the user to manually click ↺.
+// Uses the same mock infrastructure as the wizard tests above.
+
+test.describe('Correction Proposal Dialog', () => {
+  /**
+   * Add correction-proposal endpoint mocks on top of the base wizard setup.
+   * Returns a getter for the number of previewChangeSet calls made.
+   * These routes are registered AFTER setupMockAPIs so they take precedence
+   * for correction-specific endpoints.
+   */
+  const setupCorrectionMocks = async (
+    page: Page
+  ): Promise<{ getPreviewCallCount: () => number }> => {
+    let previewCallCount = 0;
+
+    // Propose a rule: one "add" op for the uncertain UNKNOWN MERCHANT transaction
+    await page.route(/corrections\.proposeChangeSet/, async (route) => {
+      const url = new URL(route.request().url());
+      const isBatch = url.searchParams.has('batch');
+      const data = {
+        changeSet: {
+          ops: [
+            {
+              op: 'add',
+              id: 'op-1',
+              data: {
+                descriptionPattern: 'UNKNOWN MERCHANT',
+                matchType: 'exact',
+                entityName: 'Woolworths',
+                entityId: 'woolworths-id',
+                transactionType: null,
+                location: null,
+                tags: [],
+              },
+            },
+          ],
+        },
+        rationale: 'Rule for UNKNOWN MERCHANT',
+        targetRules: {},
+      };
+      const responseData = { result: { data } };
+      await route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify(isBatch ? [responseData] : responseData),
+      });
+    });
+
+    // Preview: count calls, return a minimal result
+    await page.route(/corrections\.previewChangeSet/, async (route) => {
+      previewCallCount++;
+      const url = new URL(route.request().url());
+      const isBatch = url.searchParams.has('batch');
+      const data = {
+        diffs: [
+          {
+            description: 'UNKNOWN MERCHANT',
+            checksum: 'xyz789',
+            before: { entityName: null, transactionType: null, location: null },
+            after: { entityName: 'Woolworths', transactionType: null, location: null },
+            matchedRule: null,
+            status: 'matched',
+          },
+        ],
+        summary: { newMatches: 1, removedMatches: 0, statusChanges: 0 },
+      };
+      const responseData = { result: { data } };
+      await route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify(isBatch ? [responseData] : responseData),
+      });
+    });
+
+    // Apply: succeed silently
+    await page.route(/corrections\.applyChangeSet/, async (route) => {
+      const url = new URL(route.request().url());
+      const isBatch = url.searchParams.has('batch');
+      const responseData = { result: { data: { success: true } } };
+      await route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify(isBatch ? [responseData] : responseData),
+      });
+    });
+
+    return { getPreviewCallCount: () => previewCallCount };
+  };
+
+  test('dialog opens when an entity is chosen for an uncertain transaction', async ({ page }) => {
+    await setupMockAPIs(page);
+    await page.goto('/finance/import');
+    await navigateToReviewStep(page, simpleCSV);
+    await setupCorrectionMocks(page);
+
+    await page.getByRole('tab', { name: /uncertain/i }).click();
+    await page.getByRole('button', { name: /grouped/i }).click();
+
+    const group = page.getByTestId('transaction-group').first();
+    await group.getByRole('button', { name: /choose existing/i }).click();
+    await group.locator('select').selectOption('woolworths-id');
+
+    await expect(page.getByText('Correction proposal')).toBeVisible({ timeout: 8000 });
+  });
+
+  test('Apply ChangeSet is enabled after initial preview runs', async ({ page }) => {
+    await setupMockAPIs(page);
+    await page.goto('/finance/import');
+    await navigateToReviewStep(page, simpleCSV);
+    await setupCorrectionMocks(page);
+
+    await page.getByRole('tab', { name: /uncertain/i }).click();
+    await page.getByRole('button', { name: /grouped/i }).click();
+
+    const group = page.getByTestId('transaction-group').first();
+    await group.getByRole('button', { name: /choose existing/i }).click();
+    await group.locator('select').selectOption('woolworths-id');
+
+    await expect(page.getByText('Correction proposal')).toBeVisible({ timeout: 8000 });
+    const applyBtn = page.getByRole('button', { name: /Apply ChangeSet/i });
+    await expect(applyBtn).toBeEnabled({ timeout: 10000 });
+  });
+
+  test('changing Transaction type auto-reruns preview and re-enables Apply', async ({ page }) => {
+    await setupMockAPIs(page);
+    await page.goto('/finance/import');
+    await navigateToReviewStep(page, simpleCSV);
+    const { getPreviewCallCount } = await setupCorrectionMocks(page);
+
+    await page.getByRole('tab', { name: /uncertain/i }).click();
+    await page.getByRole('button', { name: /grouped/i }).click();
+
+    const group = page.getByTestId('transaction-group').first();
+    await group.getByRole('button', { name: /choose existing/i }).click();
+    await group.locator('select').selectOption('woolworths-id');
+
+    await expect(page.getByText('Correction proposal')).toBeVisible({ timeout: 8000 });
+    const applyBtn = page.getByRole('button', { name: /Apply ChangeSet/i });
+    await expect(applyBtn).toBeEnabled({ timeout: 10000 });
+
+    const callsAfterOpen = getPreviewCallCount();
+
+    // Change Transaction type — this is the bug scenario the fix addresses
+    const dialog = page.getByRole('dialog', { name: /Correction proposal/i });
+    const txnTypeSelect = dialog
+      .locator('select')
+      .filter({ has: page.locator('option[value="purchase"]') });
+    await txnTypeSelect.selectOption('purchase');
+
+    // Fix: preview must auto-rerun WITHOUT the user clicking ↺
+    await expect(async () => {
+      expect(getPreviewCallCount()).toBeGreaterThan(callsAfterOpen);
+    }).toPass({ timeout: 5000 });
+
+    await expect(applyBtn).toBeEnabled({ timeout: 5000 });
+    await expect(page.getByText(/Preview stale/i)).not.toBeVisible();
+  });
+
+  test('Apply ChangeSet closes the dialog', async ({ page }) => {
+    await setupMockAPIs(page);
+    await page.goto('/finance/import');
+    await navigateToReviewStep(page, simpleCSV);
+    await setupCorrectionMocks(page);
+
+    await page.getByRole('tab', { name: /uncertain/i }).click();
+    await page.getByRole('button', { name: /grouped/i }).click();
+
+    const group = page.getByTestId('transaction-group').first();
+    await group.getByRole('button', { name: /choose existing/i }).click();
+    await group.locator('select').selectOption('woolworths-id');
+
+    await expect(page.getByText('Correction proposal')).toBeVisible({ timeout: 8000 });
+    const applyBtn = page.getByRole('button', { name: /Apply ChangeSet/i });
+    await expect(applyBtn).toBeEnabled({ timeout: 10000 });
+    await applyBtn.click();
+
+    await expect(page.getByText('Correction proposal')).not.toBeVisible({ timeout: 5000 });
+  });
+});
